% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_serverversion.R
\name{new_serverversion}
\alias{new_serverversion}
\alias{new_serverversion}
\alias{new_serverversion_preview}
\title{new_serverversion}
\usage{
new_serverversion()

new_serverversion_preview()
}
\description{
Check if RStudio Server can be upgraded
}
\details{
Fetches the version number from
\url{https://www.rstudio.com/products/rstudio/download-server/} and
\url{https://www.rstudio.com/products/rstudio/download/preview/}.
Compares it to the local version number and throws a message if the an
upgrade can be performed.
}
\examples{
# check if your RStudio Server Preview versions can be upgraded
new_serverversion_preview()

}
