% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_server.R
\name{upgrade_server}
\alias{upgrade_server}
\alias{upgrade_server_preview}
\title{upgrade_server}
\usage{
upgrade_server_preview()
}
\description{
Upgrades your RStudio Server
}
\details{
Checks again if an upgrade is needed. If so, it downloads the newest .deb
file and installs it over the the older version. You must reload the window
in your browser and RStudio Server should inform you about the upgrade after
that.

If you perform an upgrade of RStudio Server and an existing version of the server is currently
running, then the upgrade process will also ensure that active sessions are immediately migrated to
the new version. This includes the following behavior:

\describe{
\item{}{Running R sessions are suspended so that future interactions with the server automatically launch the updated R session binary}
\item{}{Currently connected browser clients are notified that a new version is available and automatically refresh themselves.}
\item{}{The core server binary is restarted}
}
}
